Many improvements.
An initial version of the env module was added.
It provides two basic functions: merge() and purge(), which
are responsible for merging and purging variables to the
environment.
Such functionality is provided so that users may be able to specify environment
variables on a repository-specific basis. In the future, we may extend
this functionality and provide further customisation based on
environment variables.

Many docstrings were also improved. I have been reading Gentoo's Portage
source code, and am starting to follow some of their guidelines, as well
as some of Google's.

Last but not least, some minor improvements were also done to the
actions module.