A lot of improvements.
'lib/craft' has been rename to 'craft', seeing as 'lib' is not that
recommended among python projects to hold libraries.
The project's name should be used instead, so that is what I'm doing.

the load.Registry class is now being used whenever loading unit sets,
meaning that if there are name conflicts, the user will be warned about
them.